Tests refactoring: make tests more structured.
